polish code: Long -> long

+++ src/main/java/com/itranswarp/bean/SortBean.java
@@ -15,7 +15,7 @@ public class SortBean extends AbstractRequestBean {
             throw new ApiException(ApiError.PARAMETER_INVALID, "ids", "Invalid id list.");
         }
         for (Long id : ids) {
-            if (id == null || id <= 0) {
+            if (id == null || id.longValue() <= 0) {
                 throw new ApiException(ApiError.PARAMETER_INVALID, "ids", "Invalid id list.");
             }
         }
